Move attributes out of rustc_hir
