+++ compiler/rustc_attr_ir/src/stability.rs
@@ -4,7 +4,7 @@ use rustc_ast::attr::version::RustcVersion;
 use rustc_macros::{BlobDecodable, Decodable, Encodable, PrintAttribute, StableHash};
 use rustc_span::{ErrorGuaranteed, Symbol, sym};
 
-use crate::attrs::PrintAttribute;
+use crate::PrintAttribute;
 
 /// The version placeholder that recently stabilized features contain inside the
 /// `since` field of the `#[stable]` attribute.
