+++ compiler/rustc_attr_ir/src/weak_lang_items.rs
