+++ compiler/rustc_hir/src/arena.rs
@@ -4,7 +4,7 @@
 rustc_arena::declare_arena! {
     // HIR types
     asm_template: rustc_ast::InlineAsmTemplatePiece,
-    attribute: crate::Attribute,
+    attribute: rustc_attr_ir::Attribute,
     owner_info: crate::OwnerInfo<'tcx>,
     macro_def: rustc_ast::MacroDef,
     delegation_info: crate::DelegationInfo,
