+++ compiler/rustc_hir/src/lib.rs
@@ -11,35 +11,38 @@
 #![feature(derive_const)]
 #![feature(exhaustive_patterns)]
 #![feature(never_type)]
-#![feature(variant_count)]
 #![recursion_limit = "256"]
 // tidy-alphabetical-end
 
 mod arena;
-pub mod attrs;
 pub mod def;
 pub mod def_path_hash_map;
 pub mod definitions;
-pub mod diagnostic_items;
 mod hir;
 pub mod intravisit;
-pub mod lang_items;
 pub mod lints;
 pub mod pat_util;
-mod stability;
 mod stable_hash_impls;
-pub mod target;
-pub mod weak_lang_items;
+mod target_impls;
 
 #[cfg(test)]
 mod tests;
 
 #[doc(no_inline)]
 pub use hir::*;
-pub use lang_items::{LangItem, LanguageItems};
+pub use rustc_attr_ir::{self as attrs, find_attr};
 pub use rustc_hir_id::*;
 pub use rustc_span::def_id;
-pub use stability::*;
-pub use target::{MethodKind, Target};
+// FIXME: Remove this use tree, replace by `rustc_hir::attrs` or `rustc_attr_ir` imports
+#[doc(hidden)]
+pub use {
+    attrs::target::{self, MethodKind, Target},
+    attrs::{
+        AttrArgs, AttrItem, AttrPath, Attribute, ConstStability, DefaultBodyStability,
+        HashIgnoredAttrId, LangItem, LanguageItems, PartialConstStability, Stability,
+        StabilityLevel, StableSince, UnstableReason, VERSION_PLACEHOLDER,
+    },
+    attrs::{diagnostic_items, lang_items, weak_lang_items},
+};
 
 pub use crate::arena::Arena;
